add a Changes.Relation to filter out lr structs
