+++ components/electric/lib/electric/replication/changes.ex
@@ -167,4 +167,49 @@ defmodule Electric.Replication.Changes do
   def generateTag(%Transaction{origin: origin, commit_timestamp: tm}) do
     origin <> "@" <> Integer.to_string(DateTime.to_unix(tm, :millisecond))
   end
+
+  defmodule Relation.Column do
+    defstruct [:flags, :name, :type, :type_modifier, :primary_key]
+
+    @type t() :: %__MODULE__{
+            flags: [:key],
+            name: String.t(),
+            type: atom(),
+            type_modifier: integer()
+          }
+  end
+
+  defmodule Relation do
+    alias Electric.Postgres.SchemaRegistry
+    defstruct [:id, :namespace, :name, :replica_identity, :columns]
+
+    @type t() :: %__MODULE__{
+            id: Messages.relation_id(),
+            namespace: String.t(),
+            name: String.t(),
+            replica_identity: :default | :nothing | :all_columns | :index,
+            columns: [Relation.Column.t()]
+          }
+
+    # Convert a Relation message into a table structure as used by the SchemaRegistry
+    @spec to_schema_table(t()) :: {SchemaRegistry.replicated_table(), [SchemaRegistry.column()]}
+    def to_schema_table(%__MODULE__{} = relation) do
+      {%{
+         name: relation.name,
+         schema: relation.namespace,
+         oid: relation.id,
+         replica_identity: relation.replica_identity,
+         primary_keys: []
+       },
+       Enum.map(
+         relation.columns,
+         &%{
+           name: &1.name,
+           part_of_identity?: :key in &1.flags,
+           type: &1.type,
+           type_modifier: &1.type_modifier
+         }
+       )}
+    end
+  end
 end
